+++ app/types/ba_torment.go
@@ -16,8 +16,8 @@ type BATormentPartyData struct {
 type BATormentPartyDetail struct {
 	FinalRank   int              `json:"FINAL_RANK"`
 	TormentRank int              `json:"TORMENT_RANK"`
-	Score       int64            `json:"SCORE"`
-	UserID      int64            `json:"USER_ID"`
+	Score       int              `json:"SCORE"`
+	UserID      int              `json:"USER_ID"`
 	Level       string           `json:"LEVEL"`
 	PartyData   map[string][]int `json:"PARTY_DATA"`
 }
@@ -65,8 +65,8 @@ type NamedUser struct {
 // *************************************
 
 type RankData struct {
-	UserID    int64
+	UserID    int
 	FinalRank int
-	Score     int64
-	PartScore int64
+	Score     int
+	PartScore int
 }
